chore: restructured home and marketplace
