+++ PROJECT_STRUCTURE.md
@@ -1,0 +1,55 @@
+# Recommended Project Structure
+
+    1 .
+    2 ├── convex/                   // Your Convex backend functions (Good as is)
+    3 ├── public/                   // Static assets (Good as is)
+    4 ├── rental-escrow-smart-contract/ // Solidity project (Good as is)
+    5 └── src/
+    6     ├── app/                  // Next.js App Router (Routing, layouts, pages)
+    7     │   ├── (home)/           // Route group for auth pages (login, signup)
+    8     │   ├── (services)/      // Route group for marketing pages (home, about)
+    9
+
+10 │ │ ├── dashboard/
+11 │ │ ├── marketplace/
+12 │ │ └── ...
+13 │ └── api/ // API routes (Good as is)
+14 │
+15 ├── components/
+16 │ ├── ui/ // Reusable, unstyled UI primitives (shadcn/ui) (Good as is)
+17 │ └── shared/ // Shared components used across multiple features (e.g.,
+PageLayout, Navbar, Footer)
+18 │
+19 ├── features/ // <-- KEY CHANGE: Group domain logic here
+20 │ ├── home/
+21 │ │ ├── components/ // React components specific to auth
+22 │ │ ├── hooks/ // React hooks for auth
+23 │ │ └── api.ts // Auth-related API calls (e.g., login, logout)
+24 │ ├── marketplace/
+25 │ │ ├── components/ // e.g., RentalPostCard, FilterSidebar
+26 │ │ ├── hooks/ // e.g., useMarketplaceFilters
+27 │ │ └── api.ts // e.g., fetchMarketplaceListings
+28 │ ├── rental/
+29 │ │ ├── components/ // e.g., BiddingForm, RentalStatusBadge
+30 │ │ ├── hooks/
+31 │ │ └── api.ts
+32 │ └── bidding/
+33 │ ├── components/ // e.g., UserAvatar, ProfileForm
+34 │ ├── hooks/
+35 │ └── api.ts
+36 │
+37 ├── lib/ // <-- Centralized utilities and SDKs
+38 │ ├── alchemy.ts // Your Alchemy SDK initialization and core functions
+39 │ ├── convex.ts // Convex client setup
+40 │ ├── wagmi.ts // Wagmi/web3 config (Good as is)
+41 │ ├── utils.ts // General utility functions (cn, formatters, etc.) (Good as is)
+42 │ └── logger.ts // Logging utility (Good as is)
+43 │
+44 ├── providers/ // All React Context providers (Good as is, but ensure only
+providers are here)
+45 │
+46 ├── hooks/ // Truly global hooks (e.g., useMediaQuery, useLocalStorage)
+47 │
+48 ├── stores/ // Global state management (Zustand, etc.) (Good as is)
+49 │
+50 └── types/ // Global type definitions (Good as is)
