+++ src/app/(services)/(profile)/[address]/_components/nft-details-modal.tsx
