+++ src/app/(services)/marketplace/page.tsx
@@ -4,13 +4,13 @@ import { useState, useRef, Suspense } from 'react';
 import { motion, AnimatePresence, useInView } from 'framer-motion';
 import { gsap } from 'gsap';
 import { ScrollTrigger } from 'gsap/ScrollTrigger';
-import { ViewModeToggle } from '@/app/(services)/marketplace/_components/view-mode-toggle';
-import { SearchBar } from '@/app/(services)/marketplace/_components/search-bar';
-import { FilterHeader } from '@/app/(services)/marketplace/_components/filter-header';
-import { FilterSection } from '@/app/(services)/marketplace/_components/filter-section';
-import { EmptyState } from './_components/empty-state';
-import { RentalPostDetailsModal } from './_components/rental-post-details-modal';
-import { ActiveFilters } from './_components/active-filters';
+import { ViewModeToggle } from '@/features/marketplace/components/view-mode-toggle';
+import { SearchBar } from '@/features/marketplace/components/search-bar';
+import { FilterHeader } from '@/features/marketplace/components/filter-header';
+import { FilterSection } from '@/features/marketplace/components/filter-section';
+import { EmptyState } from '@/features/marketplace/components/empty-state';
+import { RentalPostDetailsModal } from '@/features/marketplace/components/rental-post-details-modal';
+import { ActiveFilters } from '@/features/marketplace/components/active-filters';
 import { useSession } from 'next-auth/react';
 import { RentalPostCardSkeleton } from '@/components/shared/skeletons/rental-post-card.skeleton';
 import { useViewMode, useSetViewMode } from '@/stores/card-view-mode.store';
