+++ src/components/shared/nft-components/attributes-section.tsx
