+++ src/components/shared/nft-components/contract-info.tsx
