+++ src/components/shared/nft-components/detail-section.tsx
