+++ src/components/shared/nft-components/image-section.tsx
