+++ src/components/shared/nft-components/media-info.tsx
