+++ src/components/shared/nft-components/nft-card.tsx
