+++ src/components/shared/nft-components/opensea-metadata.tsx
