+++ src/components/shared/nft-components/spam-warning.tsx
