+++ src/components/shared/nft-components/timeline-info.tsx
