+++ src/components/shared/nft-components/token-details.tsx
