+++ src/components/shared/nft-components/uri-section.tsx
