+++ src/features/home/components/featured-rental-posts.tsx
