+++ src/features/home/components/steps-to-rent-and-list.tsx
